refactor: App.tsx, main.tsx 파일 나누어 관심사 분리
